refactor:로또구매 클래스 패키지 변경
